+++ src/main/java/edu/online/messenger/controller/UserController.java
@@ -1,26 +1,21 @@
 package edu.online.messenger.controller;
 
-
 import edu.online.messenger.service.UserService;
 import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@AllArgsConstructor
 @RestController
 @RequestMapping("/api/users")
-
+@AllArgsConstructor
 public class UserController {
 
-    private UserService userService;
+    private final UserService userService;
 
     @GetMapping("/existence/login/{login}")
-    public ResponseEntity<Boolean> checkUserExistenceByLogin(@PathVariable String login) {
-        boolean exists = userService.userExistsByLogin(login);
-        return new ResponseEntity<>(exists, HttpStatus.OK);
+    public boolean existsByLogin(@PathVariable String login) {
+        return userService.existsByLogin(login);
     }
 }